Introducing basic rudimentary unit tests
This is the first baby step towards being able to unit test changes.
The current set of tests is just an example to start the conversation.

Researching different frameworks and methods I decided the easiest, most fitting implementation given the curent context is one where test classes are `friend` to what is to being tested - this allows access to private members etc.

Run tests with: `make tests`