Enable compiling without GUI
wxWidgets has the capability to compile with only the "base" library.
To do so, you must set the wxUSE_GUI=0 preprocessor directive when compiling.
This will disable a few things such as dragging in the some of the include
files.  So, in order to get wxDateTime, you have to manually include it.  It
no longer comes in by just including wx/wx.h.  Additionally there were a few
files in DStarRepeater that bring in DStarRepeaterApp.h that really don't
need it.