Move away from subdir build.
Building a library and then linking it doesn't work as well as just linking
the whole thing together.  Additionally, if the source needs to be compiled
with multiple flags for multiple versions of the resulting executable, this
method doesn't work so well.  We now just compile the entirety of the files
up and make it work.  This is pretty easy to do in autoconf as opposed to
having to try to manage it by hand.

This change also moves to a non-gui build for the dstarrepeaterd.