LibUSB:  Removing redundancy between Windows and Linux
This file was enclosed by a huge #if defined(WIN32) switching in an entirely
different object.  Little of the underlying code was actually different.
This pulls out only the differences and unifies the interface between the
two platforms.

NOTE:  The *actual* differences here are areally between LibUSB 0.1 and
LibUSB 1.0.  If Windows moves to LibUSB 1.0, this class can actually
become a *LOT* simpler because you can just keep the existing WIN32 way of
getting pointers to the appropriate functions, and on Linux you statically
assign the function pointers.