Fix symlinks in support files
Automake when it adds its missing support files defaults to symlinking them
to the installed autoconf versions on your systems.  That means that these
symlinks will point nowhere for a system that doesn't have autoconf installed.
The idiot packager needs to remeber to pass the --copy flag to automake when
he sets up the files so that the files will be copied and properly checked
into the repository.

Missed in #74.