Add support programs
Add the ancillary programs for the DV3000 and the StarDV to the build.  These
are controlled by the --with-stardv and --with-ambeserver flags and are not
built by default.

In addition --with-ambeserver interacts with --with-gpio, building a version
of AMBEserver that supports the wiringPi GPIO if you use --with-gpio with it.
There is a change of behavior here, as AMBEserverGPIO is never built.  If you
enable GPIO for the project, an executable named AMBEserver is produced that
includes the GPIO bits.  If you do not enable GPIO, an executable named
AMBEserver that does not include wiringPi is produced.  Only one init.d script
is installed as well.  It is called AMBEserver, and if you use the GPIO enabled
version, you will have to change the serial port name.  Ideally, this should
be settable in /etc/default/AMBEserver instead of being hard coded into the
file.