Move XReflector to GNU Autoconf
Continuation of the work to move all the projects to GNU Autoconf.

XReflector supports the --without-gui flag that the other GUI/Daemon split
apps do.  It also supports --with-dcs and --with-dextra.  These flags are
defined to be mutually exlusive and configure will not allow you to use both
of them at once as the code does not seem to support that.

With regards to DExtra support, there are quite a few files that seem to get
compiled in even if the -DDEXTRA_LINK flag is turned off.  For example,
XReflector/XReflectorDExtraHandler.cpp is referenced by
XReflector/XReflectorThread.cpp even when you don't want DExtra support.
If the code were changed we could compile XReflectorDxtraHandler.cpp and
friends only when the support was required.

Also note that ircDDBGateway has DExtra/DCS flags that I forgot to include.
Fixing this will come in another patch set.