AMBEServer:  Fix init.d install path
The init.d script for AMBEServer now has a good install path.  There are a
couple of probably needed changes here.  First, it should be moved to systemd
as nearly every distro out there has moved that way.  Second, it needs to have
the serial port settable in a more sane way as the new RPi 3 can shuffle the
serial ports some.