DStarRepeater:  Add Debian package creation directory
Add the proper files to make a package using deb-buildpkg.  This will create
three packages:

dstarrepeaterd:  The Non-GUI version of the program
dstarrepeater: The GUI version of the program
opendv-usb-tools:  The rmk8055 and rmuri programs for managing USB devices

Things should install in the normal debian places with /etc/opendv being
for configuration, /usr/share/opendv being for shared opendv files and
/var/log/opendv being for logs.  The OpenDV user is created by these installers
and the /var/log/opendv directory is created for log files.

There is also a systemd start script to start the daemons.  This should work
on systemd based stuff which is all the current things.

The program should build with --with-gpio on armhf, and the build dependencies
reflect that it will need wiringPi to work.