DStarRepeater:  Compile a common library for the common stuff
We were compiling all of the Common stuff twice, once for the GUI, and once
for the non-GUI.  This is redundant and I fixed it on the other programs, but
didn't get around to retrofitting DStarRepeater until now.  This should now
compile a static library and link it in to both of the programs, saving a lot
of compile time, especially on slower platforms.