DStarRepeater:  Change log files to be statically named
The current code rotates the log files on its own by trying to detect what
the current day is and creating new files.  This is functionality best left
for external programs on UNIX/Linux systems.  There is, in fact, logrotate
that is installed in most distributions to handle this functionality.  This way
old files can be compressed to save space and a bunch of other nifty pieces
of functionality that we get for free.  Packagers should also probably include
the logrotate config files in their packages.