DStarRepeater:  Add udev rules for USB devices
Add some udev rules to get installed on package installation that change the
permissions for USB devices to 0666.  This allows the dstarrepeater[d]
programs to access these devices without being root.

Unfortunately, under the debian package system, only a single package can
"own" a file.  This means that even though the GUI and non-GUI packages
can benefit from these rules, only one can install it and claim it.  I chose
to have the dstarrepeaterd package claim this file and install it, and then
put a dependency on dstarrepeaterd in the dstarrepeater package.  It is much
more likely that a person wants to have a completely non-gui system than the
other way around.