ircDDBGateway: Move management of common directories to the base package
Instead of having every different package check for the opendv user and the
proper directories, move this all to opendv-base and depend all the packages
on this base package.  This should make things a little cleaner at the
disadvantage of installing all the announcement files in /usr/share for
all users of any package.