DStarRepeater:  Restructure udev rule and opendv user creation
The udev rules were created by the dstarrepeaterd package, and it was made
a requirement of the GUI dstarrepeater pacakge.  This probably wasn't the way
to do it.  I created another dstarrepeater-udev package to bring these rules
in and depended both of the dstarrepeater? packages on it.  This way it stays
a little cleaner if you don't want the daemons.

Also depended dstarrepeater? on opendv-base to bring in the opendv user, log
and conf directories with proper ownership.