AMBEserver: Replace wiringPi with native calls.
WiringPi requires root privs to manipulate the GPIO pins to reset the DV3K.
The rest of AMBEserver doesn't require this privs.  There is a group on the
Raspberry Pi named "gpio" that gets permissions to all the GPIO files in the
/sys/class/gpio directory so that programs do not need root to maniuplate
these pins.  This patch rewrites the reset support so that the daemon can
run as a non-root user as part of the gpio group.  It also drops all
dependencies on wiringPi, so it's not needed as part of the build anymore nor
are there any differences between the Rapsberry Pi version of AMBEserver and
the normal versions.  One AMBEserver to rule them all!