opendv-base:  Add opendv to gpio group.
The opendv user will need to access the gpio pins on the Raspberry Pi to
support DStarRepeater, AMBEserver and friends.  This also breaks up each
group addition on a different line so that if one group add fails, the others
will still take effect.  This allows the script to work on amd64 systems that
may not have the gpio group.