AMBEserver:  Systemd changes
Make the systemd unit conflict with the getty unit and be started after it.
This will make it so that getty is actually stopped on that tty and
ambeserver will replace it.  This should do the right thing for most people.
In the case that there's no DV3000 available, upon failure of the ambserver
startup, systemd will start the ttyS0 getty.  This should cover the case
where someone is using a ThumbDV on a Raspberry Pi.