AMBEserver: Use line buffered IO for stdout/stderr
When the daemon is running under systemd, it changes stdout/stderr to use
block buffered IO.  This means that you don't get any status errors until the
program accumulates 4k of output on either of the descriptiors.  This isn't
very useful when you're using journalctl on it, so we change to line buffered
IO to try to solve the issue.