AMBEserver: Add CygWin defines to do com port numbers instead of tty.
Apparently this is intended to compile under CygWin too.  This patch puts
in #defines to make it so that the server takes a com port number with a
-c flag instead of the -i flag that takes a tty name.  This should make it
more familliar to them.  It still *eliminates* the -c usage for UNIX users as
it's not going to be right a great majority of the time and is actually
pretty confusing.