DStarRepeaterD:  Add environment variable to make wiringPi try /dev/gpiomem
WiringPi works by directly accessing /dev/mem.  This is actually a pretty bad
plan.  The kernel GPIO folks have tried to mitigate this by creating a new
device, /dev/gpiomem, that tries to mitigate some of the damage here.  Udev
sets the permissions on this device so that members of the gpio group can
access it.  opendv-base puts the opendv user in this group, so we should be
able to use this to keep from having to run DStarRepeaterD as root.  The
environment variable WIRINGPI_GPIOMEM causes wiringPi to try to use the
/dev/gpiomem method first before trying to resort to other ways.  We get
systemd to set this environment variable before trying to start DStarRepeaterD.

Note that for running DStarRepeater as a normal user with an X-Windows
interface, this environment setting will have to be an exercise for the
reader.