DStarRepeater:  Remodel UDRCController
The UDRCController was using the confusing HardwareController interface.  This
patch changes it to merely inherit off of ExternalController.  We do not use
any of the thread machinery because a GPIO call with wiringPi is merely a
register read/write.  This shouldn't take an extra thread to muck with it.
Additionally, we implement the functional modes, and dispense with writing
arbitrary pins.

Eventually there needs to be written an IExternalController interface that
can be inherited from to conform pick up all the controller types.  There can
be a common concrete helper implementation for the existing classes until they
can potentially be cleaned up and reexamined.