DStarRepeater:  Fix the failed assertions from wxShell
wxShell executes a command in the shell.  The documentation states that it
must be run from the main thread, but this isn't happening because when
the request to run a command comes from the D-STAR side, it's being executed
on the D-STAR repeater thread, rather than the main one.  This could cause
instability when this happens.  This patch installs events on the main
wxApp object and has the D-STAR repeater thread fire events so that those
commands can get executed on the main thread.  This should prevent the
assertions everyone is seeing.

There's also a bunch of cleanup here for stuff that's just messy.

Conflicts:
	DStarRepeater/DStarRepeater/DStarRepeaterApp.cpp
	DStarRepeater/DStarRepeater/DStarRepeaterApp.h
	DStarRepeater/DStarRepeater/DStarRepeaterFrame.cpp
	DStarRepeater/DStarRepeater/DStarRepeaterTRXThread.cpp