DStarRepeater:  Properly catch log file opening errors
The log file is opened in the constructor of CLogger.  If it fails, it fails
silently and allows the program to try to bootstrap itself.  This causes errors
later on and assertions file because there is no open log file.

This patch throws an exception from the constructor on file failure so that it
can be passed to CDStarRepeaterApp to be handled.  The new implementation
catches the error condition and starts to send logging output to stderr if the
log files cannot be opened.