DStarRepeater:  Fix configuration file issues
If the configuration file doesn't exist for some reason or is inaccessible
this causes assertions to fire and weird issues in the code.  This is because
the DStarRepeaterConfig object is left uninitialized.

This patch throws exceptions when the configuration file cannot be opened for
some reason, and it also provides another option to the constructor that allows
for it to except if the file doesn't exist.  File non-existence is an error
condition for the repeater itself, but not for the configuration application.
We need to change these so that the repeater can throw an error for when the
configruation file doesn't exist.