DStarRepeater:  Eliminate DStarRepeaterThreadHelper
The thread helper is completely unnecessary if all the threads just inherit
off of a DStarRepeaterThread that inherits from wxThread.  They all have a
structure that begs to just be inherited from the wxThread object.  This way
we can use standard structures to start and stop the threads and eliminate
a lot of pass-through methods on DStarRepeaterThreadHelper.

This commit makes the necessary changes to implement this in the repeater
application and the various threads.  Future changes will move more redundant
functionality into the base class.