DStarRepeater: UDRCController:  Add support for hotspots
Hotspots using the miniDIN of the UDRC will only use the PTT pin and won't
need to assert any of the other GPIO.  In addition, the sense of PTT is
reversed on the miniDIN because it's an open collector output.  This patch
inverts PTT and turns off all the additional GPIO pins so that they can be
used for other uses.