DStarRepeater:  Support for the UDRC II
The UDRC II changes the PTT pin for people using the DIN6 connector, or
"Hotspot" mode.  The pin used to be shared with the HD15 connector, but it now
shares with the HD15 pin instead.  This is so that for other applications you
can run two radios on the right and left channels of the AIC3204 and key them
separately (see SVXLink).

This patch will detect which version of the UDRC you have and change the PTT
pin automatically if you are in hotspot mode.  This change should "just work"
for users using the DIN6.